Added patientID search field and level drop down box to analyse page
The dropdown box should have the levels atempted by the patient defined
by the ID that is searched. It is not working for me atm but it may be
because I've setup the database incorrectly (the graph does not work as
well). If one of you could try it and lmk if it functions correctly,
that would be great