Added progress graphs
Based on the patientID it plots time of all attempts against their
success rate (four graphs added in total - there are coin/wall options
for both left and right progress).
Also in the analyse page, depending on which graph option you choose,
the fields with unnecessary info become greyed out so that the user does
not get confused.